reorganizing evalutor code and renaming of modules
in preparation for future updates to the mpi code, all evaluator code has been reorganized. All evaluators go into their respective module, and code sthared among 2 or more goes into a new util package.

All relevant modules also have been renamed to future_{style of parallelization}
