--- docs/source/ema_documentation/em_framework/ema_ipyparallel.rst
@@ -1,6 +1,0 @@
-**********************
-:mod:`ema_ipyparallel`
-**********************
-
-.. automodule:: ema_workbench.em_framework.ema_ipyparallel
-   :members:
