--- docs/source/ema_documentation/em_framework/ema_multiprocessing.rst
@@ -1,6 +1,0 @@
-**************************
-:mod:`ema_multiprocessing`
-**************************
-
-.. automodule:: ema_workbench.em_framework.ema_multiprocessing
-   :members:
