+++ docs/source/ema_documentation/em_framework/futures_ipyparallel.rst
@@ -1,0 +1,6 @@
+**************************
+:mod:`futures_ipyparallel`
+**************************
+
+.. automodule:: ema_workbench.em_framework.futures_ipyparallel
+   :members:
