+++ docs/source/ema_documentation/em_framework/futures_mpi.rst
@@ -1,0 +1,6 @@
+******************
+:mod:`futures_mpi`
+******************
+
+.. automodule:: ema_workbench.em_framework.futures_mpi
+   :members:
