+++ docs/source/ema_documentation/em_framework/futures_multiprocessing.rst
@@ -1,0 +1,6 @@
+******************************
+:mod:`futures_multiprocessing`
+******************************
+
+.. automodule:: ema_workbench.em_framework.futures_multiprocessing
+   :members:
