+++ docs/source/ema_documentation/em_framework/futures_util.rst
@@ -1,0 +1,6 @@
+******************
+:mod:`futures_util`
+******************
+
+.. automodule:: ema_workbench.em_framework.futures_util
+   :members:
