+++ ema_workbench/em_framework/__init__.py
@@ -1,7 +1,6 @@
 import warnings
 
 __all__ = [
-    "ema_parallel",
     "parameters",
     "model",
     "outcomes",
@@ -77,8 +76,6 @@ from .salib_samplers import SobolSampler, MorrisSampler, FASTSampler, get_SALib_
 from .evaluators import (
     perform_experiments,
     optimize,
-    MPIEvaluator,
-    MultiprocessingEvaluator,
     SequentialEvaluator,
     Samplers,
 )
@@ -96,6 +93,9 @@ from .optimization import (
     to_problem,
     to_robust_problem,
 )
+from .futures_ipyparallel import IpyparallelEvaluator
+from .futures_multiprocessing import MultiprocessingEvaluator
+from .futures_mpi import MPIEvaluator
 from .outputspace_exploration import OutputSpaceExploration
 
 try:
