--- test/test_em_framework/test_ema_multiprocessing.py
@@ -1,11 +1,0 @@
-"""
-
-
-"""
-
-
-# Created on 14 Mar 2017
-#
-# .. codeauthor::jhkwakkel <j.h.kwakkel (at) tudelft (dot) nl>
-
-__all__ = []
